Update fisherman.maryreyes.md
He hasn't been available for a formal interview. I would need to talk to him for half an hour at the very list to have enough information to fill in all of the gaps. I did go out on the boat with him to film and take pictures, so I have visuals for the story. Let me know if you have any questions. :)