Individual story
As I’ve worked further on this, I’ve decided that it’s going to be more of a photo story. Jana did not want me to use the formal name of the UF group that she practices with for liability reasons. Therefore, I am focusing more on her doing it recreationally on campus and pursuing partner work for the first time. I will be doing a follow-up interview on Tuesday as well as taking photos. I can include more details after that interview.