Update Aerial Silk
I know this is kind of a photo story so I'm not that worried about your transitions.
Just some suggestions:
Where does she get all the energy to do this? Might be worth a follow up interview if she didn't say to tie it in with the theme.
How old is Joseph? 
I'm not crazy about stacking quotes on top of each other, especially if they are from different people. So I paraphrased one quote from Rosen to break up some of the double decker quotes.