Update bhagan.md
Hi Viana! 

This is a really interesting topic. I actually wrote an article on this topic when I took reporting. Here are some of the things I noticed: 

I think you should do a list format for the myths so that you don’t have to repeat the first sentence of each paragraph. Instead of “Myth number one is that… Myth number two is that,” etc. you can just write the myths next to the numbers (e.g. 1. Cramming all night long for an exam will help improve grades). That could also help the designers highlight the main points of the article during the design process. 

The ending also feels a bit disjointed from the rest of the article. I would try to add a sentence or paragraph that connects the last myth to Somers’ experience. Maybe something like, “Some students learn the hard way…” 

If possible, I would try to speak to a health professional about this and maybe include a quote from them. That would make it seem more reliable. When I wrote about this topic, I called a sleep clinic in Gainesville, and one of the ladies there was super helpful and extremely willing to talk to me about this. Just call a few places and see if anyone is available. If not, the article would still be fine without it. 

You also made a lot of minor grammatical errors. I highlighted them for you in the Word document I used to edit. Leave your email in the comments, and I’ll send it to you. Let me know if you have any questions. 

Thanks! 
Mary