Update self-defense.md
Hi Katina, 
That was amazing! I only spotted one minor mistake in your whole article. That makes you the Muhammad Ali of this class. You just kicked ass.

The error I found was in the paragraph that began with "Since becoming state director..." You included a "the" that doesn't belong, and you should add an "s" to "area."

Other than that, this article was the epitome of perfection. The topic was not only interesting, but you talked to multiple sources that enhanced the story. It also hits home for me because one of my best friends was raped in high school. People like her need to find ways to feel empowered again, and I think R.A.D. classes are a good place to start.

Let me know if you have any questions.

Thanks!
Mary