Edited
As I'm reading this it's 8 a.m. and I just made a cup of coffee....haha.

Anyway, this is really interesting and has good evidence to support it. If you didn't talk to the neuroscientist personally, maybe be more specific about where you found the information instead of just using the word "research." Something that I think could make it stronger is interviewing a local professional about the topic--maybe a nutritionist or a professor from the College of Food Science and Human Nutrition. I didn't find many technical errors because you're a rockstar at that stuff. Let's talk more in class about your idea for the infographic!